# api/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Partial_User_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_UserRoles_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"User": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"password",
					"name",
					"email",
					"id"
				],
				"type": "object",
				"description": "Model User"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/user-permission/get-all": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/Partial_User_"
											},
											{
												"properties": {
													"userRoles": {
														"items": {
															"$ref": "#/components/schemas/Partial_UserRoles_"
														},
														"type": "array"
													}
												},
												"required": [
													"userRoles"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user-permission/create": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"type": "number",
										"format": "double"
									},
									"password": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"password",
									"name",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/update/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userRoles": {
										"items": {
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"id"
											],
											"type": "object"
										},
										"type": "array"
									},
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"userRoles",
									"name",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/delete/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/server-status": {
			"get": {
				"operationId": "GetServerStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"serverTime": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"serverTime",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/server-status/routes": {
			"get": {
				"operationId": "GetServerRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}